+++ modules/apim-adapter/src/main/java/com/axway/apim/adapter/user/APIManagerUserAdapter.java
@@ -26,7 +26,7 @@ import com.axway.apim.adapter.APIManagerAdapter.CacheType;
 import com.axway.apim.api.model.Image;
 import com.axway.apim.api.model.Organization;
 import com.axway.apim.api.model.User;
-import com.axway.apim.lib.CommandParameters;
+import com.axway.apim.lib.CoreParameters;
 import com.axway.apim.lib.errorHandling.AppException;
 import com.axway.apim.lib.errorHandling.ErrorCode;
 import com.axway.apim.lib.utils.rest.DELRequest;
@@ -43,7 +43,7 @@ import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
 
 public class APIManagerUserAdapter {
 	
-	CommandParameters cmd = CommandParameters.getInstance();
+	CoreParameters cmd = CoreParameters.getInstance();
 	
 	private static Logger LOG = LoggerFactory.getLogger(APIManagerUserAdapter.class);
 	
@@ -74,7 +74,7 @@ public class APIManagerUserAdapter {
 		URI uri;
 		HttpResponse httpResponse = null;
 		try {
-			uri = new URIBuilder(CommandParameters.getInstance().getAPIManagerURL()).setPath(RestAPICall.API_VERSION + "/users"+userId)
+			uri = new URIBuilder(CoreParameters.getInstance().getAPIManagerURL()).setPath(RestAPICall.API_VERSION + "/users"+userId)
 					.addParameters(filter.getFilters())
 					.build();
 			RestAPICall getRequest = new GETRequest(uri, APIManagerAdapter.hasAdminAccount());