+++ modules/apis/src/main/java/com/axway/apim/api/export/lib/ChangeAPICLIOptions.java
@@ -4,6 +4,8 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.apache.commons.cli.ParseException;
 
+import com.axway.apim.lib.errorHandling.AppException;
+
 public class ChangeAPICLIOptions extends APIExportCLIOptions {
 
 	CommandLine cmd;
@@ -42,5 +44,11 @@ public class ChangeAPICLIOptions extends APIExportCLIOptions {
 		return "Application-Export";
 	}
 
-
+	public APIChangeParams getAPIChangeParams() throws AppException {
+		APIChangeParams params = new APIChangeParams();
+		super.addAPIExportParams(params);
+		params.setNewBackend(getValue("newBackend"));
+		params.setOldBackend(getValue("oldBackend"));
+		return params;
+	}
 }