+++ modules/apis/src/main/java/com/axway/apim/apiimport/lib/APIImportParams.java
@@ -1,38 +1,61 @@
 package com.axway.apim.apiimport.lib;
 
-import com.axway.apim.lib.APIMCoreCLIOptions;
-import com.axway.apim.lib.CommandParameters;
-import com.axway.apim.lib.EnvironmentProperties;
-import com.axway.apim.lib.errorHandling.AppException;
+import com.axway.apim.lib.CoreParameters;
+import com.axway.apim.lib.StandardImportParams;
 
-public class APIImportParams extends CommandParameters {
-
-	public APIImportParams(APIMCoreCLIOptions parser)
-			throws AppException {
-		super(parser.getCmd(), parser.getInternalCmd(), new EnvironmentProperties(parser.getCmd().getOptionValue("stage"), parser.getCmd().getOptionValue("swaggerPromoteHome")));
-	}
+public class APIImportParams extends StandardImportParams {
+	
+	private Boolean forceUpdate;
+	private Boolean useFEAPIDefinition;
+	
+	private Boolean changeOrganization = false; 
+	
+	private String apiDefintion;
 	
 	public static synchronized APIImportParams getInstance() {
-		return (APIImportParams)CommandParameters.getInstance();
+		return (APIImportParams)CoreParameters.getInstance();
 	}
 	
 	@Override
-	public boolean ignoreCache() {
+	public boolean isIgnoreCache() {
 		// For import action we ignore the cache in all cases!
 		return true;
 	}
-	
-	public boolean isUseFEAPIDefinition() {
-		if(getValue("useFEAPIDefinition")==null) return false;
-		return Boolean.parseBoolean(getValue("useFEAPIDefinition"));
+
+	public Boolean isUseFEAPIDefinition() {
+		if(useFEAPIDefinition==null) return false;
+		return useFEAPIDefinition;
 	}
-	
-	public boolean isForceUpdate() {
-		try {
-			if(hasOption("forceUpdate")) {
-				return true;
-			} 
-		} catch (Exception ignore) {}
-		return false;
+
+	public void setUseFEAPIDefinition(Boolean useFEAPIDefinition) {
+		if(useFEAPIDefinition==null) return;
+		this.useFEAPIDefinition = useFEAPIDefinition;
+	}
+
+	public Boolean isForceUpdate() {
+		if(forceUpdate==null) return false;
+		return forceUpdate;
+	}
+
+	public void setForceUpdate(Boolean forceUpdate) {
+		if(forceUpdate==null) return;
+		this.forceUpdate = forceUpdate;
+	}
+
+	public String getApiDefintion() {
+		return apiDefintion;
+	}
+
+	public void setApiDefintion(String apiDefintion) {
+		this.apiDefintion = apiDefintion;
+	}
+
+	public Boolean isChangeOrganization() {
+		return changeOrganization;
+	}
+
+	public void setChangeOrganization(Boolean changeOrganization) {
+		if(changeOrganization==null) return;
+		this.changeOrganization = changeOrganization;
 	}
 }