+++ modules/apps/src/main/java/com/axway/apim/appimport/lib/AppImportParams.java
@@ -1,23 +1,16 @@
 package com.axway.apim.appimport.lib;
 
-import com.axway.apim.lib.APIMCoreCLIOptions;
-import com.axway.apim.lib.CommandParameters;
-import com.axway.apim.lib.EnvironmentProperties;
-import com.axway.apim.lib.errorHandling.AppException;
+import com.axway.apim.lib.CoreParameters;
+import com.axway.apim.lib.StandardImportParams;
 
-public class AppImportParams extends CommandParameters {
-
-	public AppImportParams(APIMCoreCLIOptions parser)
-			throws AppException {
-		super(parser.getCmd(), parser.getInternalCmd(), new EnvironmentProperties(parser.getCmd().getOptionValue("stage"), parser.getCmd().getOptionValue("swaggerPromoteHome")));
-	}
+public class AppImportParams extends StandardImportParams {
 	
 	public static synchronized AppImportParams getInstance() {
-		return (AppImportParams)CommandParameters.getInstance();
+		return (AppImportParams)CoreParameters.getInstance();
 	}
 	
 	@Override
-	public boolean ignoreCache() {
+	public boolean isIgnoreCache() {
 		// For import action we ignore the cache in all cases!
 		return true;
 	}